refactor(lotto): moved User class to the lotto package from the user package
Moved the User class from the user package to the lotto package to better organize related classes. Added a List<Lotto> constructor in the User class for improved functionality.
